==> 01_preprocess_reverse_n_inout.ipynb <==
import os, sys
import glob

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import statsmodels.api as sm

import datetime
import time

# df_db의 reverse 방향 노선 생성

df_db = pd.read_csv(f"./final_db/df_db_hday.csv", index_col=[0])
df_db = df_db.reindex(columns=['holiday', 'date', 'time', 'line', 'order', 'from_stt', 'to_stt'])
df_db['order']='상선'


df_rvs = df_db.copy()
df_rvs.columns = ['holiday', 'date', 'time', 'line', 'order', 'from_stt', 'to_stt']
df_rvs = df_rvs.reindex(columns=['holiday', 'date', 'time', 'line', 'order', 'to_stt', 'from_stt'])
df_rvs.columns = ['holiday', 'date', 'time', 'line', 'order', 'from_stt', 'to_stt']
df_rvs['order'] = '하선'

df_db = df_db.append(df_rvs).reset_index(drop=True)
df_db['in'] = np.nan
df_db['out'] = np.nan
df_db['crowd'] = np.nan

df_db = df_db.drop_duplicates()
df_db = df_db.reset_index(drop=True)
df_db.to_csv(f"./final_db/df_db_hday_rvs.csv")

df_db

# df_inout 생성

# inout 데이터 새로 가공
df_inout = pd.read_csv(f"./database/Subway_day_hour/2015_subway_day_hour.csv", encoding='cp949', index_col=[0]).reset_index()
df_inout['역명'] = df_inout['역명'].str.replace(r"\([^)]*\)", "", regex=True)
for col in df_inout.columns[4:]:
    df_inout.loc[: , col] = df_inout.loc[: , col].str.replace(",", "")
df_inout['line'] = df_inout['역번호']//100
df_inout = df_inout.drop(columns = ["역번호"])
#df_inout.loc[df_inout['역명'] == '까치산', 'line'] = 2


df_inout.columns = ['date', 'station', 'inout', '5', '6', '7', '8', '9', '10', '11', '12', 
              '13', '14', '15', '16', '17', '18', '19', '20', '21', '22', '23', '24', 'line']
df_inout = df_inout.reindex(columns = ['date', 'line', 'station', 'inout', '5', '6', '7', '8', '9', '10', '11', '12', 
                           '13', '14', '15', '16', '17', '18', '19', '20', '21', '22', '23', '24',])

# 1, 2, 3, 4호선만 남기기
df_inout = df_inout.loc[df_inout['line'].apply(lambda x: x in [1, 2, 3, 4])]
df_inout = df_inout.sort_values(by=["date", "line"]).reset_index(drop=True)

# 2호선 내선/외선 이외의 분기 노선 모두 삭제 
line2_ex_list = ['까치산', '신정네거리', '양천구청', '도림천', '용답', '신답', '용두', '신설동']
cond = df_inout.loc[:, 'station'].apply(lambda x: x in line2_ex_list)

drop_idx = df_inout.loc[(df_inout['line']==2) & cond, :].index
df_inout = df_inout.drop(index=drop_idx)

df_inout = df_inout.drop_duplicates()
df_inout = df_inout.reset_index(drop=True)

# inout 데이터 새로 가공
df_tmp = pd.read_csv(f"./database/Subway_day_hour/2016_subway_day_hour.csv", encoding='cp949', index_col=[0]).reset_index()
df_tmp['역명'] = df_tmp['역명'].str.replace(r"\([^)]*\)", "", regex=True)
for col in df_tmp.columns[4:]:
    df_tmp.loc[: , col] = df_tmp.loc[: , col].str.replace(",", "")
df_tmp['line'] = df_tmp['역번호']//100
df_tmp = df_tmp.drop(columns = ["역번호"])
#df_tmp.loc[df_tmp['역명'] == '까치산', 'line'] = 2


df_tmp.columns = ['date', 'station', 'inout', '5', '6', '7', '8', '9', '10', '11', '12', 
              '13', '14', '15', '16', '17', '18', '19', '20', '21', '22', '23', '24', 'line']
df_tmp = df_tmp.reindex(columns = ['date', 'line', 'station', 'inout', '5', '6', '7', '8', '9', '10', '11', '12', 
                           '13', '14', '15', '16', '17', '18', '19', '20', '21', '22', '23', '24',])

# 1, 2, 3, 4호선만 남기기
df_tmp = df_tmp.loc[df_tmp['line'].apply(lambda x: x in [1, 2, 3, 4])]
df_tmp = df_tmp.sort_values(by=["date", "line"]).reset_index(drop=True)

# 2호선 내선/외선 이외의 분기 노선 모두 삭제 
line2_ex_list = ['까치산', '신정네거리', '양천구청', '도림천', '용답', '신답', '용두', '신설동']
cond = df_tmp.loc[:, 'station'].apply(lambda x: x in line2_ex_list)

drop_idx = df_tmp.loc[(df_tmp['line']==2) & cond, :].index
df_tmp = df_tmp.drop(index=drop_idx)

df_tmp = df_tmp.drop_duplicates()
df_tmp = df_tmp.reset_index(drop=True)

# 전처리: na 값들 해결 -> 2016 년 데이터에는 결측치가 없다! 그러므로 여기서 가져오장! 
# (df_inout.groupby('station').sum() - df_tmp.groupby('station').sum()).mean()
# 결과가 -5.675676이므로 그냥 가져오자. 몰랑........ㅎ
st_t = time.time()
print(f"Start {time.time()-st_t}")

for idx, row in df_inout.loc[df_inout['5'].isna()].iterrows():
    #idx, row['date'], row['line'], row['station'], row['inout'], row['5']
    cond1 = df_tmp['date'].str[-5:] == row['date'][-5:]
    cond2 = (df_tmp['station'] == row['station'])& (df_tmp['line'] == row['line'])
    cond3 = df_tmp['inout'] == row['inout']
    cond = cond1&cond2&cond3
    
    df_inout.loc[idx, '5'] = df_tmp.loc[cond,'5'].values[0]
    
print(f"End {time.time()-st_t}")


# 전처리! inout에 있는 신천 -> 잠실새내
df_inout['station'] = df_inout['station'].replace("신천", "잠실새내")


df_keep = df_inout.copy()
df_keep = df_keep.drop_duplicates()
df_keep = df_keep.reset_index(drop=True)
df_keep.to_csv("./final_db/2015_subway_day_hour_1234_wo_ch3.csv")

df_inout = pd.read_csv("./final_db/2015_subway_day_hour_1234_wo_ch3.csv", index_col=[0])

# 3호선 충무로 승차, 하차 row append 하기.
df_inout = df_inout.append(pd.DataFrame({'date': [date for date in df_inout['date'].unique()],
                            'line': 3, 'station': '충무로', 'inout': '승차'
                           }))

df_inout = df_inout.append(pd.DataFrame({'date': [date for date in df_inout['date'].unique()],
                            'line': 3, 'station': '충무로', 'inout': '하차'
                           }))

df_inout = df_inout.reset_index(drop=True)

line3_list = ['금호', '약수', '동대입구', '충무로', '을지로3가', '종로3가', '안국']

st_t = time.time()
for date in df_inout['date'].unique():
    cond1 = (df_inout['date']==date)&(df_inout['line']==3)
    cond2 = df_inout['station'].isin(line3_list)
    
    cond3 = df_inout['inout'] == '승차'
    cond4 = df_inout['inout'] == '하차'
    
    cond5 = df_inout['station'] == '충무로'
    
    df_3ch_in = df_inout.loc[cond1&cond2&cond3].set_index('station').reindex(line3_list)
    df_3ch_in = df_3ch_in.interpolate(axis=0)
    df_inout.loc[cond1&cond3&cond5, '5':] = df_3ch_in.loc['충무로', '5':].to_numpy()
    
    df_3ch_out = df_inout.loc[cond1&cond2&cond3].set_index('station').reindex(line3_list)
    df_3ch_out = df_3ch_out.interpolate(axis=0)
    df_inout.loc[cond1&cond4&cond5, '5':] = df_3ch_out.loc['충무로', '5':].to_numpy()
print(time.time()-st_t)

df_inout = df_inout.drop_duplicates()
df_inout = df_inout.reset_index(drop=True)
df_inout.to_csv("./final_db/2015_subway_day_hour_1234.csv")

df_inout
