==> 02_process_concat_base_inout.ipynb <==
import os, sys
import glob

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import statsmodels.api as sm

import datetime
import time

# df_db와 df_inout 연결

df_db = pd.read_csv(f"./final_db/df_db_hday_rvs.csv", index_col=[0])
df_inout = pd.read_csv("./final_db/2015_subway_day_hour_1234.csv", index_col=[0])

df_inout.columns = ['date', 'line', 'stt', 'inout', ] + list(df_inout.columns)[-20:]
df_inout['line'] = df_inout['line'].apply(lambda x: f"{x}호선")
df_inout


df_inout만 연결해보자. : 1호선 작동하고 쥬겄어용....

df_db = pd.read_csv(f"./final_db/df_db_hday_rvs.csv", index_col=[0])
df_db = df_db.drop_duplicates()

df_db1 = pd.read_csv(f"./final_db/df_db_hday_rvs_inout_line_1.csv", index_col = 0)
df_db1 = df_db1.drop_duplicates()
df_db1 = df_db1.reset_index(drop=True)

df_db.loc[df_db['line'] == '1호선', ['in', 'out']] = df_db1.loc[df_db1['line'] == '1호선', ['in', 'out']].copy()

# df_db2 = pd.read_csv(f"./final_db/df_db_hday_rvs_inout_line_2.csv", index_col = 0)
# df_db2 = df_db2.drop_duplicates()
# df_db2 = df_db2.reset_index(drop=True)

# df_db.loc[df_db['line'] == '2호선', ['in', 'out']] = df_db2.loc[df_db2['line'] == '2호선', ['in', 'out']].copy()


3, 4호선은 왜..? 

st_t = time.time()
for line in df_db['line'].unique():
    print(line)
    
    if line in ['1호선', ]:#'2호선']:
        continue

    cond_db1 = df_db['line']==line
    cond_inout1 = df_inout['line']==line
    
    for date in df_db['date'].unique():
        cond_db2 = df_db['date']==date
        cond_inout2 = df_inout['date']==date
        
        for stt in df_db.loc[cond_db1, 'from_stt'].unique():
            cond_db3 = df_db['from_stt']== stt
            cond_inout3 = df_inout['stt']== stt
            
            cond_in = df_inout['inout'] == '승차'
            cond_out = df_inout['inout'] == '하차'

            copy_in = df_inout.loc[cond_inout1&cond_inout2&cond_inout3&cond_in, '5':].to_numpy().reshape(-1,1)
            copy_out = df_inout.loc[cond_inout1&cond_inout2&cond_inout3&cond_out, '5':].to_numpy().reshape(-1,1)
            
            try:
                df_db.loc[cond_db1&cond_db2&cond_db3, 'in'] = copy_in
                df_db.loc[cond_db1&cond_db2&cond_db3, 'out'] = copy_out
            except:
                try:
                    copy_in = np.append(copy_in, copy_in, axis=0)
                    copy_out = np.append(copy_out, copy_out, axis=0)

                    df_db.loc[cond_db1&cond_db2&cond_db3, 'in'] = copy_in
                    df_db.loc[cond_db1&cond_db2&cond_db3, 'out'] = copy_out
                except:
                    print(line, date, stt)
                    df_err = df_db.loc[cond_db1&cond_db2&cond_db3, :].copy()
                    print("$"*30)
                    print(df_inout.loc[cond_inout1&cond_inout2&cond_inout3&cond_in, '5':])
                    break
                
        #break
        print(f"Date {date} end: {time.time() - st_t}")
    #break
    print(f"Line no.{line} end: {time.time() - st_t}")

df_db

df_db.to_csv(f"./final_db/df_db_hday_rvs_inout.csv")
