==> 03_add_answer_crowdness.ipynb <==
import os, sys
import glob

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import statsmodels.api as sm

import datetime
import time

# df_db와 df_inout, df_crd 연결

df_db = pd.read_csv(f"./final_db/df_db_hday_rvs.csv", index_col=[0])
#df_inout = pd.read_csv("./final_db/2015_subway_day_hour_1234.csv", index_col=[0])
df_crd = pd.read_csv('./database/crowded2015.csv', index_col=[0])

df_db = df_db.drop_duplicates()
df_crd = df_crd.drop_duplicates()

df_crd = df_crd.groupby(by=['Day', 'Line', 'Direction', 'Station']).mean()
df_crd = df_crd.reset_index(drop=False)

# 2호선 내선/외선 이외의 분기 노선 모두 삭제 
line2_ex_list = ['까치산', '신정네거리', '양천구청', '도림천', '용답', '신답', '용두', '신설동']
cond = df_crd.loc[:, 'Station'].apply(lambda x: x in line2_ex_list)

drop_idx = df_crd.loc[(df_crd['Line']=='2호선') & cond, :].index
df_crd = df_crd.drop(index=drop_idx)
df_crd = df_crd.reset_index(drop=True)

df_crd.columns = ['holiday', 'line', 'order', 'stt', ] + list(df_crd.columns)[-20:]

def holiday(x):
    if x == "평일":
        return 0
    elif x == "토요일":
        return 1
    else:
        return 2
    
df_crd['holiday'] = df_crd['holiday'].apply(lambda x: holiday(x))
df_crd


st_t = time.time()
for line in df_db['line'].unique():
    cond_db1 = df_db['line']==line
    cond_crd1 = df_crd['line']==line
    
    for order in df_db['order'].unique():
        cond_db2 = df_db['order']==order
        cond_crd2 = df_crd['order']==order
        
        for holiday in df_db['holiday'].unique():
            cond_db3 = df_db['holiday']== holiday
            cond_crd3 = df_crd['holiday']== holiday
            
            for stt in df_db.loc[cond_db1&cond_db2, 'from_stt'].unique():
                cond_db4 = df_db['from_stt']== stt
                cond_crd4 = df_crd['stt'] == stt
                
                try:
                    N = (df_db.loc[cond_db1&cond_db2&cond_db3&cond_db4, :].shape[0])//20
                    print(df_db.loc[cond_db1&cond_db2&cond_db3&cond_db4, :].shape)
                    copy = df_crd.loc[cond_crd1&cond_crd2&cond_crd3&cond_crd4, '5':].to_numpy().reshape(-1,1)
                    
                    print(copy.shape, N)
                    copy = np.vstack([copy]*N)
                    df_db.loc[cond_db1&cond_db2&cond_db3&cond_db4, "crowd"] = copy
                except:
                    print(df_db.loc[cond_db1&cond_db2&cond_db3&cond_db4, :])
                    print(df_crd.loc[cond_crd1&cond_crd2&cond_crd3&cond_crd4,:])
                    print(copy, N)
                 
        print(f"Order {order} end: {time.time() - st_t}")
    print(f"Line no.{line} end: {time.time() - st_t}")

df_db.to_csv(f"./final_db/df_db_hday_rvs_crd.csv")

df_db

crd, inout 파일 합치기

df_db_inout = pd.read_csv(f"./final_db/df_db_hday_rvs_inout.csv", index_col=[0])
df_db_inout = df_db_inout.drop(columns=['cong', 'crowd'])

df_db_inout

df_db_crd = df_db.copy()
df_db_crd = df_db_crd.drop(columns=['in', 'out'])

df_db_crd

df_db_inout.merge(df_db_crd, how='outer').to_csv("./final_db/data.csv")

df_db_inout.merge(df_db_crd, how='outer')
