==> 04_process_weather.ipynb <==
import os, sys
import glob

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import statsmodels.api as sm

import datetime

# preprocess weather 정보
# df_weather = pd.DataFrame()
# for fname in os.listdir("./weather_hour/"):
#     df_tmp = pd.read_csv(f"./weather_hour/{fname}", encoding="cp949")
#     if df_weather.empty:
#         df_weather = df_tmp.copy()
#     else:
#         df_weather = df_weather.append(df_tmp)
# df_weather = df_weather.reset_index(drop=True)
# df_weather.loc[:, '일시'] = df_weather.loc[:, '일시'].apply(lambda x: datetime.datetime.strptime(x, "%Y-%m-%d %H:%M"))
# df_weather.to_csv("./database/weather_all.csv")

# fillna
- 날씨마루에서 제공하는 결측치 보간 방법 사용
- https://bd.kma.go.kr/kma2020/dta/edu/KBP57200_Python.do

df_weather = pd.read_csv("./database/weather_all.csv", index_col=[0])

cond = ~(df_weather.isna().sum(axis=1) == 0)
df_weather.loc[cond]

# '강수량(mm)', '적설(cm)' 는 0으로 치환
df_weather['강수량(mm)'] = df_weather['강수량(mm)'].fillna(value = 0)
df_weather['적설(cm)'] = df_weather['강수량(mm)'].fillna(value = 0)

# '현지기압(hPa)', 는 ffill, bfill
df_weather['현지기압(hPa)'] = df_weather['현지기압(hPa)'].fillna(method='ffill')
df_weather['현지기압(hPa)'] = df_weather['현지기압(hPa)'].fillna(method='bfill')

for col in ['기온(°C)', '풍속(m/s)', '습도(%)', '일조(hr)', '전운량(10분위)']:
    df_weather[col] = df_weather[col].fillna(method='ffill')
    df_weather[col] = df_weather[col].fillna(method='bfill')

df_weather.loc[cond]

df_weather.to_csv("./final_db/df_weather.csv")

전체 db랑 합쳐주자

import datetime

df_data = pd.read_csv("./final_db/data.csv", index_col=[0])

for weather_col in list(df_weather.columns[3:]):
    df_data[weather_col] = np.nan

drop_24_idx = df_data.loc[df_data['time']==24].index

df_data = df_data.drop(index=drop_24_idx)

for date in df_data['date'].unique():
    for time in df_data['time'].unique():
        time = int(time)
        cond = (df_data['date']==date) & (df_data['time']== time)
        
        when = datetime.datetime.strptime(f"{date} {time}:00:00", "%Y-%m-%d %H:%M:%S")
        
        df_tmp = df_weather.loc[df_weather['일시']==str(when), '기온(°C)':].reset_index(drop=True)
        df_data.loc[cond, '기온(°C)':] = df_tmp

df_data_weather = df_data.copy()

df_data_weather.fillna(method='ffill')

df_data_weather.to_csv("./final_db/data_w_weather.csv")

# utils: 앞뒤값 살펴보기
na_col = '기온(°C)'
# ['지점', '지점명', '일시', '기온(°C)', '강수량(mm)', '풍속(m/s)', 
# '습도(%)', '현지기압(hPa)', '일조(hr)', '적설(cm)', '전운량(10분위)']

na_idx = df_weather.loc[df_weather[na_col].isna()].index
fna_idx = []

for idx in na_idx:
    fna_idx += [ i for i in range(idx-3,idx+3)]

fna_idx = sorted(list(set(fna_idx)))
fna_idx = [x for x in fna_idx if x>=0 and x<= df_weather.index.max()]
df_weather.iloc[fna_idx, :]

# 기상청에서 이렇게 하더만 안되겠구만... 그냥 interpolate 하자.. 
# import sklearn.neighbors._base
# sys.modules['sklearn.neighbors.base'] = sklearn.neighbors._base

# from missingpy import KNNImputer


# print(df_weather[na_col].isnull().sum())
# temp_idx = df_weather.loc[df_weather[na_col].isnull(), :].index
# df_weather.loc[df_weather[na_col].isnull(), :]


# # '기온(°C)', '풍속(m/s)', '습도(%)' 
# # KnnImputation을 활용하여 결측치 처리
# imputer = KNNImputer(n_neighbors = 20)
# df_weather[na_col] = pd.DataFrame(imputer.fit_transform(df_weather[na_col].to_numpy().reshape(-1, 1)))
# df_weather[na_col].isnull().sum()
# df_weather.iloc[temp_idx, :]
